% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EstimateModel.R
\name{EstimateModel}
\alias{EstimateModel}
\title{Estimate model of pair trade
Assumption is that difference between log(price) of two stocks is random walk}
\usage{
EstimateModel(price.pair, method = lm)
}
\arguments{
\item{price.pair}{A data frame with two columns of prices, one for each stock}

\item{method}{method of estimation. usually lm}
}
\value{
list of mean, std. dev.,spread, hedge ratio, and premium
}
\description{
Estimate model of pair trade
Assumption is that difference between log(price) of two stocks is random walk
}
\examples{
x    <- rnorm(1000)+50
y    <- (x - 2) + rnorm(1000)+50
price.pair.history<-as.data.frame(cbind(x,y))
model <- EstimateModel(price.pair.history, method = lm)
}

