% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pairFIX.R
\docType{package}
\name{pairFIX}
\alias{pairFIX}
\alias{pairFIX-package}
\title{pairFIX: A package for Pairs Trading via FIX engine}
\description{
This package implements a simple pair trading model and also allows the user
to execute trades via the FIX engine. The user can download stock price data and
portolio data from their broker via FIX, can estimate a pairs trading model,
make trading recommendations based on this model, and then execute trades based
on these recommendations via the FIX engine.
}

