% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EstimateModel.R
\name{RecommendTrade}
\alias{RecommendTrade}
\title{Determine if difference between stocks is stationary and if trade is recommended
Uses two tests: Phillips-Perron test for unit root, and Augmented Dickey Fuller test
If residual between two stationary stocks is more than threshold standard deviations away from mean,
then recommend trade. Otherwise return NA}
\usage{
RecommendTrade(model, todaysPrices, threshold)
}
\arguments{
\item{model}{model list returned by EstimateModel}

\item{todaysPrices}{data frame of todays prices for the stocks in model}

\item{threshold}{std deviation threshold for trade}
}
\value{
stock to buy if trade is recommended, or NA if not
}
\description{
Determine if difference between stocks is stationary and if trade is recommended
Uses two tests: Phillips-Perron test for unit root, and Augmented Dickey Fuller test
If residual between two stationary stocks is more than threshold standard deviations away from mean,
then recommend trade. Otherwise return NA
}
\examples{
x    <- rnorm(1000)+50
y    <- (x - 2) + rnorm(1000)+50
price.pair.history<-as.data.frame(cbind(x,y))
model <- EstimateModel(price.pair.history, method = lm)
x    <- rnorm(1)
y    <- (x - 2) + rnorm(1)
price.pair.today<-as.data.frame(cbind(x,y))

trade=RecommendTrade(model,price.pair.today,0.05)
}

