% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{stock.prices.hist}
\alias{stock.prices.hist}
\title{Stock prices for 4 stocks for the last year}
\format{An xts object with 244 rows and 4 variables:
\describe{
  \item{AAPL.Adjusted}{AAPL daily adjusted price}
  \item{GOOG.Adjusted}{GOOG daily adjusted price}
  \item{MSFT.Adjusted}{MSFT daily adjusted price}
  \item{AMZN.Adjusted}{AMZN daily adjusted price}
}}
\source{
quantmod package
}
\usage{
stock.prices.hist
}
\description{
AAPL, AMZN, MSFT, GOOG
}
\keyword{datasets}

